requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.22.0
SQLAlchemy>=2.0
alembic==1.11.1
asyncpg==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
